# backend/gmail_agent_openapi.yaml
openapi: 3.0.0
info:
  title: Gmail Agent API
  description: Agent for managing Gmail emails with list, read, send, and search capabilities
  version: 1.0.0
  contact:
    name: HackTheAgent Team

servers:
  - url: http://localhost:8000
    description: Local development server

paths:
  /gmail/list:
    get:
      summary: List emails from Gmail inbox
      description: Retrieve a list of emails with optional pagination and filtering
      operationId: listEmails
      x-openai-isConsequential: false
      tags:
        - Gmail
      parameters:
        - name: max_results
          in: query
          description: Maximum number of emails to return
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: page_token
          in: query
          description: Token for pagination to get next page of results
          required: false
          schema:
            type: string
        - name: query
          in: query
          description: Gmail search query (e.g., 'from:example@gmail.com', 'subject:meeting', 'is:unread')
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved emails
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  emails:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailSummary'
                  next_page_token:
                    type: string
                    nullable: true
                  result_size_estimate:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /gmail/read/{email_id}:
    get:
      summary: Read email details
      description: Get full details of a specific email by its ID
      operationId: readEmailDetails
      x-openai-isConsequential: false
      tags:
        - Gmail
      parameters:
        - name: email_id
          in: path
          description: Gmail message ID
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successfully retrieved email details
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  email:
                    $ref: '#/components/schemas/EmailDetails'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /gmail/send:
    post:
      summary: Send email
      description: Send a new email via Gmail
      operationId: sendEmail
      x-openai-isConsequential: true
      tags:
        - Gmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendEmailRequest'
      responses:
        '200':
          description: Email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message_id:
                    type: string
                  thread_id:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /gmail/search:
    post:
      summary: Search emails
      description: Search emails with advanced filters
      operationId: searchEmails
      x-openai-isConsequential: false
      tags:
        - Gmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchEmailsRequest'
      responses:
        '200':
          description: Search completed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  emails:
                    type: array
                    items:
                      $ref: '#/components/schemas/EmailSummary'
                  next_page_token:
                    type: string
                    nullable: true
                  result_size_estimate:
                    type: integer
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /gmail/register:
    post:
      summary: Register Gmail agent
      description: Register the Gmail agent with Watson Orchestrate
      operationId: registerGmailAgent
      x-openai-isConsequential: true
      tags:
        - Admin
      responses:
        '200':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent_id:
                    type: string
                  name:
                    type: string
                  status:
                    type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  schemas:
    EmailSummary:
      type: object
      properties:
        id:
          type: string
          description: Gmail message ID
        thread_id:
          type: string
          description: Gmail thread ID
        subject:
          type: string
          description: Email subject
        from:
          type: string
          description: Sender email address
        to:
          type: string
          description: Recipient email address
        date:
          type: string
          description: Email date
        snippet:
          type: string
          description: Email preview snippet
        labels:
          type: array
          items:
            type: string
          description: Gmail labels

    EmailDetails:
      allOf:
        - $ref: '#/components/schemas/EmailSummary'
        - type: object
          properties:
            body:
              type: string
              description: Full email body content

    SendEmailRequest:
      type: object
      required:
        - to
        - subject
        - body
      properties:
        to:
          type: string
          description: Recipient email address
          example: recipient@example.com
        subject:
          type: string
          description: Email subject
          example: Meeting Invitation
        body:
          type: string
          description: Email body content
          example: Hi, let's schedule a meeting for next week.
        cc:
          type: string
          description: CC recipients (comma-separated)
          example: cc1@example.com,cc2@example.com
        bcc:
          type: string
          description: BCC recipients (comma-separated)
          example: bcc@example.com
        html:
          type: boolean
          description: Whether body is HTML formatted
          default: false

    SearchEmailsRequest:
      type: object
      properties:
        from_email:
          type: string
          description: Filter by sender email
          example: sender@example.com
        to_email:
          type: string
          description: Filter by recipient email
          example: recipient@example.com
        subject:
          type: string
          description: Filter by subject keywords
          example: meeting
        after_date:
          type: string
          description: Filter emails after date (YYYY/MM/DD format)
          example: "2026/01/01"
        before_date:
          type: string
          description: Filter emails before date (YYYY/MM/DD format)
          example: "2026/12/31"
        has_attachment:
          type: boolean
          description: Filter emails with attachments
        is_unread:
          type: boolean
          description: Filter unread emails
        max_results:
          type: integer
          description: Maximum results to return
          default: 10
          minimum: 1
          maximum: 100

    Error:
      type: object
      properties:
        detail:
          type: string
          description: Error message